acmicpc/src/test/java/problem/step02/Q14681Test.java
package problem.step02;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;

class Q14681Test {

    @ParameterizedTest(name = "{index} [{arguments}]")
    @MethodSource
    void test(int x, int y, int expected) throws Exception {
        assertThat(Q14681.solution(x, y)).isEqualTo(expected);
    }

    private static Stream<Arguments> test() {
        return Stream.of(
                Arguments.of(12, 5, 1),
                Arguments.of(9, -13, 4)
        );
    }
}